@@ -3,17 +3,30 @@
 // www.ebenmonney.com/templates
 // =============================
 
+using DAL;
+using DAL.Core;
+using DAL.Core.Interfaces;
+using DAL.Models;
+using IdentityServer4.AccessTokenValidation;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
+using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Identity;
+using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
+using Microsoft.IdentityModel.Logging;
+using Microsoft.OpenApi.Models;
+using QuickApp.Authorization;
+using QuickApp.Helpers;
 using System;
 using System.Collections.Generic;
-using System.Linq;
+using System.Reflection;
 using System.Threading.Tasks;
-using DAL;
-using QuickApp.Helpers;
+using AppPermissions = DAL.Core.ApplicationPermissions;
 
 namespace QuickApp
 {
@@ -21,11 +34,206 @@ namespace QuickApp
     {
         public static void Main(string[] args)
         {
-            var host = CreateHostBuilder(args).Build();
+            var builder = WebApplication.CreateBuilder(args);
+            AddServices(builder);// Add services to the container.
+
+            var app = builder.Build();
+            ConfigureRequestPipeline(app); // Configure the HTTP request pipeline.
+
+            SeedDatabase(app); //Seed initial database
+
+            app.Run();
+        }
 
 
-            //Seed database
-            using (var scope = host.Services.CreateScope())
+        private static void AddServices(WebApplicationBuilder builder)
+        {
+            var connectionString = builder.Configuration.GetConnectionString("DefaultConnection") ??
+                            throw new InvalidOperationException("Connection string 'DefaultConnection' not found.");
+
+            var authServerUrl = builder.Configuration["AuthServerUrl"].TrimEnd('/');
+
+            string migrationsAssembly = typeof(Program).GetTypeInfo().Assembly.GetName().Name; //QuickApp
+
+            builder.Services.AddDbContext<ApplicationDbContext>(options =>
+                options.UseSqlServer(connectionString, b => b.MigrationsAssembly(migrationsAssembly)));
+
+            // add identity
+            builder.Services.AddIdentity<ApplicationUser, ApplicationRole>()
+                .AddEntityFrameworkStores<ApplicationDbContext>()
+                .AddDefaultTokenProviders();
+
+            // Configure Identity options and password complexity here
+            builder.Services.Configure<IdentityOptions>(options =>
+            {
+                // User settings
+                options.User.RequireUniqueEmail = true;
+
+                //// Password settings
+                //options.Password.RequireDigit = true;
+                //options.Password.RequiredLength = 8;
+                //options.Password.RequireNonAlphanumeric = false;
+                //options.Password.RequireUppercase = true;
+                //options.Password.RequireLowercase = false;
+
+                //// Lockout settings
+                //options.Lockout.DefaultLockoutTimeSpan = TimeSpan.FromMinutes(30);
+                //options.Lockout.MaxFailedAccessAttempts = 10;
+            });
+
+            // Adds IdentityServer.
+            builder.Services.AddIdentityServer(o =>
+            {
+                o.IssuerUri = authServerUrl;
+            })
+              // The AddDeveloperSigningCredential extension creates temporary key material for signing tokens.
+              // This might be useful to get started, but needs to be replaced by some persistent key material for production scenarios.
+              // See http://docs.identityserver.io/en/release/topics/crypto.html#refcrypto for more information.
+              .AddDeveloperSigningCredential()
+              .AddInMemoryPersistedGrants()
+              // To configure IdentityServer to use EntityFramework (EF) as the storage mechanism for configuration data (rather than using the in-memory implementations),
+              // see https://identityserver4.readthedocs.io/en/release/quickstarts/8_entity_framework.html
+              .AddInMemoryIdentityResources(IdentityServerConfig.GetIdentityResources())
+              .AddInMemoryApiScopes(IdentityServerConfig.GetApiScopes())
+              .AddInMemoryApiResources(IdentityServerConfig.GetApiResources())
+              .AddInMemoryClients(IdentityServerConfig.GetClients())
+              .AddAspNetIdentity<ApplicationUser>()
+              .AddProfileService<ProfileService>();
+
+            builder.Services.AddAuthentication(o =>
+            {
+                o.DefaultScheme = IdentityServerAuthenticationDefaults.AuthenticationScheme;
+                o.DefaultAuthenticateScheme = IdentityServerAuthenticationDefaults.AuthenticationScheme;
+                o.DefaultChallengeScheme = IdentityServerAuthenticationDefaults.AuthenticationScheme;
+            })
+               .AddIdentityServerAuthentication(options =>
+               {
+                   options.Authority = authServerUrl;
+                   options.RequireHttpsMetadata = false; // Note: Set to true in production
+                   options.ApiName = IdentityServerConfig.ApiName;
+               });
+
+            builder.Services.AddAuthorization(options =>
+            {
+                options.AddPolicy(Authorization.Policies.ViewAllUsersPolicy, policy => policy.RequireClaim(ClaimConstants.Permission, AppPermissions.ViewUsers));
+                options.AddPolicy(Authorization.Policies.ManageAllUsersPolicy, policy => policy.RequireClaim(ClaimConstants.Permission, AppPermissions.ManageUsers));
+
+                options.AddPolicy(Authorization.Policies.ViewAllRolesPolicy, policy => policy.RequireClaim(ClaimConstants.Permission, AppPermissions.ViewRoles));
+                options.AddPolicy(Authorization.Policies.ViewRoleByRoleNamePolicy, policy => policy.Requirements.Add(new ViewRoleAuthorizationRequirement()));
+                options.AddPolicy(Authorization.Policies.ManageAllRolesPolicy, policy => policy.RequireClaim(ClaimConstants.Permission, AppPermissions.ManageRoles));
+
+                options.AddPolicy(Authorization.Policies.AssignAllowedRolesPolicy, policy => policy.Requirements.Add(new AssignRolesAuthorizationRequirement()));
+            });
+
+            // Add cors
+            builder.Services.AddCors();
+
+            builder.Services.AddControllersWithViews();
+
+            builder.Services.AddSwaggerGen(c =>
+            {
+                c.SwaggerDoc("v1", new OpenApiInfo { Title = IdentityServerConfig.ApiFriendlyName, Version = "v1" });
+                c.OperationFilter<AuthorizeCheckOperationFilter>();
+                c.AddSecurityDefinition("oauth2", new OpenApiSecurityScheme
+                {
+                    Type = SecuritySchemeType.OAuth2,
+                    Flows = new OpenApiOAuthFlows
+                    {
+                        Password = new OpenApiOAuthFlow
+                        {
+                            TokenUrl = new Uri("/connect/token", UriKind.Relative),
+                            Scopes = new Dictionary<string, string>()
+                            {
+                                { IdentityServerConfig.ApiName, IdentityServerConfig.ApiFriendlyName }
+                            }
+                        }
+                    }
+                });
+            });
+
+            builder.Services.AddAutoMapper(typeof(Program));
+
+            // Configurations
+            builder.Services.Configure<AppSettings>(builder.Configuration);
+
+            // Business Services
+            builder.Services.AddScoped<IEmailSender, EmailSender>();
+
+            // Repositories
+            builder.Services.AddScoped<IUnitOfWork, HttpUnitOfWork>();
+            builder.Services.AddScoped<IAccountManager, AccountManager>();
+
+            // Auth Handlers
+            builder.Services.AddSingleton<IAuthorizationHandler, ViewUserAuthorizationHandler>();
+            builder.Services.AddSingleton<IAuthorizationHandler, ManageUserAuthorizationHandler>();
+            builder.Services.AddSingleton<IAuthorizationHandler, ViewRoleAuthorizationHandler>();
+            builder.Services.AddSingleton<IAuthorizationHandler, AssignRolesAuthorizationHandler>();
+
+            // DB Creation and Seeding
+            builder.Services.AddTransient<IDatabaseInitializer, DatabaseInitializer>();
+
+            //File Logger
+            builder.Logging.AddFile(builder.Configuration.GetSection("Logging"));
+
+            //Email Templates
+            EmailTemplates.Initialize(builder.Environment);
+        }
+
+
+        private static void ConfigureRequestPipeline(WebApplication app)
+        {
+            if (app.Environment.IsDevelopment())
+            {
+                app.UseDeveloperExceptionPage();
+                IdentityModelEventSource.ShowPII = true;
+            }
+            else
+            {
+                app.UseExceptionHandler("/Error");
+                // The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
+                app.UseHsts();
+            }
+
+            app.UseHttpsRedirection();
+            app.UseStaticFiles();
+            app.UseRouting();
+            app.UseCors(builder => builder
+                .AllowAnyOrigin()
+                .AllowAnyHeader()
+                .AllowAnyMethod());
+
+            app.UseIdentityServer();
+            app.UseAuthorization();
+
+            app.UseSwagger();
+            app.UseSwaggerUI(c =>
+            {
+                c.DocumentTitle = "Swagger UI - QuickApp";
+                c.SwaggerEndpoint("/swagger/v1/swagger.json", $"{IdentityServerConfig.ApiFriendlyName} V1");
+                c.OAuthClientId(IdentityServerConfig.SwaggerClientID);
+                c.OAuthClientSecret("no_password"); //Leaving it blank doesn't work
+            });
+
+            app.UseEndpoints(endpoints =>
+            {
+                endpoints.MapControllerRoute(
+                    name: "default",
+                    pattern: "{controller}/{action=Index}/{id?}");
+
+                endpoints.Map("api/{**slug}", context =>
+                {
+                    context.Response.StatusCode = StatusCodes.Status404NotFound;
+                    return Task.CompletedTask;
+                });
+
+                endpoints.MapFallbackToFile("index.html");
+            });
+        }
+
+
+        private static void SeedDatabase(WebApplication app)
+        {
+            using (var scope = app.Services.CreateScope())
             {
                 var services = scope.ServiceProvider;
 
@@ -42,25 +250,6 @@ namespace QuickApp
                     throw new Exception(LoggingEvents.INIT_DATABASE.Name, ex);
                 }
             }
-
-            host.Run();
         }
-
-
-        public static IHostBuilder CreateHostBuilder(string[] args) =>
-            Host.CreateDefaultBuilder(args)
-                .ConfigureWebHostDefaults(webBuilder =>
-                {
-                    webBuilder.UseStartup<Startup>();
-                })
-                .ConfigureLogging((hostingContext, logging) =>
-                {
-                    logging.ClearProviders();
-                    logging.AddConfiguration(hostingContext.Configuration.GetSection("Logging"));
-                    logging.AddConsole();
-                    logging.AddDebug();
-                    logging.AddEventSourceLogger();
-                    logging.AddFile(hostingContext.Configuration.GetSection("Logging"));
-                });
     }
 }
