@@ -9,7 +9,8 @@ import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-banner-demo',
-  templateUrl: './banner-demo.component.html'
+  templateUrl: './banner-demo.component.html',
+  styleUrls: ['./banner-demo.component.scss']
 })
 export class BannerDemoComponent {
 
