@@ -3,13 +3,11 @@
 // www.ebenmonney.com/templates
 // =============================
 
-import { Component, OnInit, OnDestroy } from '@angular/core';
+import { Component, OnInit, OnDestroy, ViewChild } from '@angular/core';
 import { AlertService, DialogType, MessageSeverity } from '../../services/alert.service';
-import { Subscription, Observable, fromEvent, of, merge } from 'rxjs';
-import { map, distinctUntilChanged } from 'rxjs/operators';
+import { BaseChartDirective } from 'ng2-charts';
 
 require('chart.js');
-
 
 
 @Component({
@@ -19,91 +17,113 @@ require('chart.js');
 })
 export class StatisticsDemoComponent implements OnInit, OnDestroy {
 
-  chartData = [
-    { data: [65, 59, 80, 81, 56, 55], label: 'Series A' },
-    { data: [28, 48, 40, 19, 86, 27], label: 'Series B' },
-    { data: [18, 48, 77, 9, 100, 27], label: 'Series C' }
-  ];
+  chartOptions;
+  chartType = 'line';
   chartLabels = ['Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun'];
-  chartOptions = {
-    responsive: true,
-    title: {
-      display: false,
-      fontSize: 16,
-      text: 'Important Stuff'
-    }
-  };
-  chartColors = [
-    { // grey
+  chartData = [
+    {
+      data: [65, 59, 80, 81, 56, 55],
+      label: 'Series A',
+      fill: 'origin',
+      /*
+      // grey color
       backgroundColor: 'rgba(148,159,177,0.2)',
       borderColor: 'rgba(148,159,177,1)',
       pointBackgroundColor: 'rgba(148,159,177,1)',
       pointBorderColor: '#fff',
       pointHoverBackgroundColor: '#fff',
       pointHoverBorderColor: 'rgba(148,159,177,0.8)'
+      */
     },
-    { // dark grey
+    {
+      data: [28, 48, 40, 19, 86, 27],
+      label: 'Series B',
+      fill: 'origin',
+      /*
+      // dark grey color
       backgroundColor: 'rgba(77,83,96,0.2)',
       borderColor: 'rgba(77,83,96,1)',
       pointBackgroundColor: 'rgba(77,83,96,1)',
       pointBorderColor: '#fff',
       pointHoverBackgroundColor: '#fff',
       pointHoverBorderColor: 'rgba(77,83,96,1)'
+      */
     },
-    { // something else
+    {
+      data: [18, 48, 77, 9, 100, 27],
+      label: 'Series C',
+      fill: 'origin',
+      /*
+      // darker grey color
       backgroundColor: 'rgba(128,128,128,0.2)',
       borderColor: 'rgba(128,128,128,1)',
       pointBackgroundColor: 'rgba(128,128,128,1)',
       pointBorderColor: '#fff',
       pointHoverBackgroundColor: '#fff',
       pointHoverBorderColor: 'rgba(128,128,128,0.8)'
+      */
     }
   ];
-  chartLegend = true;
-  chartType = 'line';
 
   timerReference: any;
-  windowWidth$: Observable<number>;
-  windowWidthSub: Subscription;
 
+
+  @ViewChild(BaseChartDirective)
+  chart?: BaseChartDirective;
 
   constructor(private alertService: AlertService) {
 
   }
 
-
   ngOnInit() {
+    this.refreshChartOptions();
     this.timerReference = setInterval(() => this.randomize(), 5000);
-
-    const initialWidth$ = of(window.innerWidth);
-    const resizedWidth$ = fromEvent(window, 'resize').pipe(map((event: any) => event.target.innerWidth as number));
-    this.windowWidth$ = merge(initialWidth$, resizedWidth$).pipe(distinctUntilChanged());
-
-    this.windowWidthSub = this.windowWidth$.subscribe(width => this.chartLegend = width < 600 ? false : true);
   }
 
   ngOnDestroy() {
     clearInterval(this.timerReference);
-    this.windowWidthSub.unsubscribe();
   }
 
-
-
-  randomize(): void {
-    const chartData = new Array(this.chartData.length);
-    for (let i = 0; i < this.chartData.length; i++) {
-      chartData[i] = { data: new Array(this.chartData[i].data.length), label: this.chartData[i].label };
-
-      for (let j = 0; j < this.chartData[i].data.length; j++) {
-        chartData[i].data[j] = Math.floor((Math.random() * 100) + 1);
+  refreshChartOptions() {
+    const baseOptions = {
+      responsive: true,
+      maintainAspectRatio: false,
+      title: {
+        display: false,
+        fontSize: 16,
+        text: 'Important Stuff'
       }
     }
 
-    this.chartData = chartData;
+    if (this.chartType != 'line') {
+      this.chartOptions = baseOptions;
+    }
+    else {
+      const lineChartOptions = {
+        elements: {
+          line: {
+            tension: 0.5
+          }
+        }
+      }
+
+      this.chartOptions = { ...baseOptions, ...lineChartOptions };
+    }
+  }
+
+  randomize(): void {
+    for (let i = 0; i < this.chartData.length; i++) {
+      for (let j = 0; j < this.chartData[i].data.length; j++) {
+        this.chartData[i].data[j] = Math.floor((Math.random() * 100) + 1);
+      }
+    }
+
+    this.chart?.update();
   }
 
   changeChartType(type: string) {
     this.chartType = type;
+    this.refreshChartOptions();
   }
 
   showMessage(msg: string): void {
@@ -131,10 +151,10 @@ export class StatisticsDemoComponent implements OnInit, OnDestroy {
   }
 
   chartClicked(e): void {
-    console.log(e);
+    //console.log(e);
   }
 
   chartHovered(e): void {
-    console.log(e);
+    //console.log(e);
   }
 }
