@@ -7,7 +7,7 @@ import { Directive, ElementRef, Output, EventEmitter, OnDestroy, NgZone } from '
 import { Subscription, fromEvent } from 'rxjs';
 
 
-declare var $: any;
+declare var bootstrap: any;
 
 interface EventArg { type: string; target: Element; relatedTarget: Element; }
 
@@ -30,12 +30,12 @@ export class BootstrapTabDirective implements OnDestroy {
 
   constructor(private el: ElementRef, private zone: NgZone) {
 
-    this.tabShownSubscription = fromEvent($(this.el.nativeElement), 'show.bs.tab')
+    this.tabShownSubscription = fromEvent(this.el.nativeElement, 'show.bs.tab')
       .subscribe((e: any) => {
         this.runInZone(() => this.showBSTab.emit({ type: e.type, target: e.target, relatedTarget: e.relatedTarget }));
       });
 
-    this.tabHiddenSubscription = fromEvent($(this.el.nativeElement), 'hidden.bs.tab')
+    this.tabHiddenSubscription = fromEvent(this.el.nativeElement, 'hidden.bs.tab')
       .subscribe((e: any) => {
         this.runInZone(() => this.hideBSTab.emit({ type: e.type, target: e.target, relatedTarget: e.relatedTarget }));
       });
@@ -56,6 +56,7 @@ export class BootstrapTabDirective implements OnDestroy {
 
 
   show(selector: string) {
-    $(selector).tab('show');
+    const tab = bootstrap.Tab.getOrCreateInstance(selector)
+    tab.show()
   }
 }
