@@ -14,237 +14,268 @@ import { Utilities } from './utilities';
 import { environment } from '../../environments/environment';
 
 interface UserConfiguration {
-    language: string;
-    homeUrl: string;
-    themeId: number;
-    showDashboardStatistics: boolean;
-    showDashboardNotifications: boolean;
-    showDashboardTodo: boolean;
-    showDashboardBanner: boolean;
+  language: string;
+  homeUrl: string;
+  themeId: number;
+  showDashboardStatistics: boolean;
+  showDashboardNotifications: boolean;
+  showDashboardTodo: boolean;
+  showDashboardBanner: boolean;
 }
 
 @Injectable()
 export class ConfigurationService {
 
-    constructor(
-        private localStorage: LocalStoreManager,
-        private translationService: AppTranslationService,
-        private themeManager: ThemeManager) {
+  constructor(
+    private localStorage: LocalStoreManager,
+    private translationService: AppTranslationService,
+    private themeManager: ThemeManager) {
 
-        this.loadLocalChanges();
+    this.loadLocalChanges();
+  }
+
+  set language(value: string) {
+    this._language = value;
+    this.saveToLocalStore(value, DBkeys.LANGUAGE);
+    this.translationService.changeLanguage(value);
+  }
+  get language() {
+    return this._language || ConfigurationService.defaultLanguage;
+  }
+
+  set themeId(value: number) {
+    value = +value;
+    this._themeId = value;
+    this.saveToLocalStore(value, DBkeys.THEME_ID);
+    this.themeManager.installTheme(this.themeManager.getThemeByID(value));
+  }
+  get themeId() {
+    return this._themeId || ConfigurationService.defaultThemeId;
+  }
+
+  set homeUrl(value: string) {
+    this._homeUrl = value;
+    this.saveToLocalStore(value, DBkeys.HOME_URL);
+  }
+  get homeUrl() {
+    return this._homeUrl || ConfigurationService.defaultHomeUrl;
+  }
+
+  set showDashboardStatistics(value: boolean) {
+    this._showDashboardStatistics = value;
+    this.saveToLocalStore(value, DBkeys.SHOW_DASHBOARD_STATISTICS);
+  }
+  get showDashboardStatistics() {
+    return this._showDashboardStatistics != null ? this._showDashboardStatistics : ConfigurationService.defaultShowDashboardStatistics;
+  }
+
+  set showDashboardNotifications(value: boolean) {
+    this._showDashboardNotifications = value;
+    this.saveToLocalStore(value, DBkeys.SHOW_DASHBOARD_NOTIFICATIONS);
+  }
+  get showDashboardNotifications() {
+    return this._showDashboardNotifications != null ? this._showDashboardNotifications : ConfigurationService.defaultShowDashboardNotifications;
+  }
+
+  set showDashboardTodo(value: boolean) {
+    this._showDashboardTodo = value;
+    this.saveToLocalStore(value, DBkeys.SHOW_DASHBOARD_TODO);
+  }
+  get showDashboardTodo() {
+    return this._showDashboardTodo != null ? this._showDashboardTodo : ConfigurationService.defaultShowDashboardTodo;
+  }
+
+  set showDashboardBanner(value: boolean) {
+    this._showDashboardBanner = value;
+    this.saveToLocalStore(value, DBkeys.SHOW_DASHBOARD_BANNER);
+  }
+  get showDashboardBanner() {
+    return this._showDashboardBanner != null ? this._showDashboardBanner : ConfigurationService.defaultShowDashboardBanner;
+  }
+  public static readonly appVersion: string = '6.0.0';
+
+  // ***Specify default configurations here***
+  public static readonly defaultLanguage: string = 'en';
+  public static readonly defaultHomeUrl: string = '/';
+  public static readonly defaultThemeId: number = 1;
+  public static readonly defaultShowDashboardStatistics: boolean = true;
+  public static readonly defaultShowDashboardNotifications: boolean = true;
+  public static readonly defaultShowDashboardTodo: boolean = false;
+  public static readonly defaultShowDashboardBanner: boolean = true;
+
+  public baseUrl = environment.baseUrl || Utilities.baseUrl();
+  public loginUrl = environment.loginUrl;
+  public fallbackBaseUrl = 'https://quickapp-standard.ebenmonney.com';
+  // ***End of defaults***
+
+  private _language: string = null;
+  private _homeUrl: string = null;
+  private _themeId: number = null;
+  private _showDashboardStatistics: boolean = null;
+  private _showDashboardNotifications: boolean = null;
+  private _showDashboardTodo: boolean = null;
+  private _showDashboardBanner: boolean = null;
+  private onConfigurationImported: Subject<void> = new Subject<void>();
+
+  configurationImported$ = this.onConfigurationImported.asObservable();
+
+  private loadLocalChanges() {
+    if (this.localStorage.exists(DBkeys.LANGUAGE)) {
+      this._language = this.localStorage.getDataObject<string>(DBkeys.LANGUAGE);
+      this.translationService.changeLanguage(this._language);
+    } else {
+      this.resetLanguage();
     }
 
-    set language(value: string) {
-        this._language = value;
-        this.saveToLocalStore(value, DBkeys.LANGUAGE);
-        this.translationService.changeLanguage(value);
-    }
-    get language() {
-        return this._language || ConfigurationService.defaultLanguage;
+    if (this.localStorage.exists(DBkeys.THEME_ID)) {
+      this._themeId = this.localStorage.getDataObject<number>(DBkeys.THEME_ID);
+      this.themeManager.installTheme(this.themeManager.getThemeByID(this._themeId));
+    } else {
+      this.resetTheme();
     }
 
-    set themeId(value: number) {
-        value = +value;
-        this._themeId = value;
-        this.saveToLocalStore(value, DBkeys.THEME_ID);
-        this.themeManager.installTheme(this.themeManager.getThemeByID(value));
-    }
-    get themeId() {
-        return this._themeId || ConfigurationService.defaultThemeId;
+    if (this.localStorage.exists(DBkeys.HOME_URL)) {
+      this._homeUrl = this.localStorage.getDataObject<string>(DBkeys.HOME_URL);
     }
 
-    set homeUrl(value: string) {
-        this._homeUrl = value;
-        this.saveToLocalStore(value, DBkeys.HOME_URL);
-    }
-    get homeUrl() {
-        return this._homeUrl || ConfigurationService.defaultHomeUrl;
+    if (this.localStorage.exists(DBkeys.SHOW_DASHBOARD_STATISTICS)) {
+      this._showDashboardStatistics = this.localStorage.getDataObject<boolean>(DBkeys.SHOW_DASHBOARD_STATISTICS);
     }
 
-    set showDashboardStatistics(value: boolean) {
-        this._showDashboardStatistics = value;
-        this.saveToLocalStore(value, DBkeys.SHOW_DASHBOARD_STATISTICS);
-    }
-    get showDashboardStatistics() {
-        return this._showDashboardStatistics != null ? this._showDashboardStatistics : ConfigurationService.defaultShowDashboardStatistics;
+    if (this.localStorage.exists(DBkeys.SHOW_DASHBOARD_NOTIFICATIONS)) {
+      this._showDashboardNotifications = this.localStorage.getDataObject<boolean>(DBkeys.SHOW_DASHBOARD_NOTIFICATIONS);
     }
 
-    set showDashboardNotifications(value: boolean) {
-        this._showDashboardNotifications = value;
-        this.saveToLocalStore(value, DBkeys.SHOW_DASHBOARD_NOTIFICATIONS);
-    }
-    get showDashboardNotifications() {
-        return this._showDashboardNotifications != null ? this._showDashboardNotifications : ConfigurationService.defaultShowDashboardNotifications;
+    if (this.localStorage.exists(DBkeys.SHOW_DASHBOARD_TODO)) {
+      this._showDashboardTodo = this.localStorage.getDataObject<boolean>(DBkeys.SHOW_DASHBOARD_TODO);
     }
 
-    set showDashboardTodo(value: boolean) {
-        this._showDashboardTodo = value;
-        this.saveToLocalStore(value, DBkeys.SHOW_DASHBOARD_TODO);
+    if (this.localStorage.exists(DBkeys.SHOW_DASHBOARD_BANNER)) {
+      this._showDashboardBanner = this.localStorage.getDataObject<boolean>(DBkeys.SHOW_DASHBOARD_BANNER);
     }
-    get showDashboardTodo() {
-        return this._showDashboardTodo != null ? this._showDashboardTodo : ConfigurationService.defaultShowDashboardTodo;
-    }
+  }
 
-    set showDashboardBanner(value: boolean) {
-        this._showDashboardBanner = value;
-        this.saveToLocalStore(value, DBkeys.SHOW_DASHBOARD_BANNER);
-    }
-    get showDashboardBanner() {
-        return this._showDashboardBanner != null ? this._showDashboardBanner : ConfigurationService.defaultShowDashboardBanner;
-    }
-    public static readonly appVersion: string = '5.0.0';
+  private saveToLocalStore(data: any, key: string) {
+    setTimeout(() => this.localStorage.savePermanentData(data, key));
+  }
 
-    // ***Specify default configurations here***
-    public static readonly defaultLanguage: string = 'en';
-    public static readonly defaultHomeUrl: string = '/';
-    public static readonly defaultThemeId: number = 1;
-    public static readonly defaultShowDashboardStatistics: boolean = true;
-    public static readonly defaultShowDashboardNotifications: boolean = true;
-    public static readonly defaultShowDashboardTodo: boolean = false;
-    public static readonly defaultShowDashboardBanner: boolean = true;
+  public import(jsonValue: string) {
+    this.clearLocalChanges();
 
-    public baseUrl = environment.baseUrl || Utilities.baseUrl();
-    public tokenUrl = environment.tokenUrl || environment.baseUrl || Utilities.baseUrl();
-    public loginUrl = environment.loginUrl;
-    public fallbackBaseUrl = 'https://quickapp.ebenmonney.com';
-    // ***End of defaults***
+    if (jsonValue) {
+      const importValue: UserConfiguration = Utilities.JsonTryParse(jsonValue);
 
-    private _language: string = null;
-    private _homeUrl: string = null;
-    private _themeId: number = null;
-    private _showDashboardStatistics: boolean = null;
-    private _showDashboardNotifications: boolean = null;
-    private _showDashboardTodo: boolean = null;
-    private _showDashboardBanner: boolean = null;
-    private onConfigurationImported: Subject<boolean> = new Subject<boolean>();
+      if (importValue.language != null) {
+        this.language = importValue.language;
+      }
 
-    configurationImported$ = this.onConfigurationImported.asObservable();
+      if (importValue.themeId != null) {
+        this.themeId = +importValue.themeId;
+      }
 
-    private loadLocalChanges() {
-        if (this.localStorage.exists(DBkeys.LANGUAGE)) {
-            this._language = this.localStorage.getDataObject<string>(DBkeys.LANGUAGE);
-            this.translationService.changeLanguage(this._language);
-        } else {
-            this.resetLanguage();
-        }
+      if (importValue.homeUrl != null) {
+        this.homeUrl = importValue.homeUrl;
+      }
 
-        if (this.localStorage.exists(DBkeys.THEME_ID)) {
-            this._themeId = this.localStorage.getDataObject<number>(DBkeys.THEME_ID);
-            this.themeManager.installTheme(this.themeManager.getThemeByID(this._themeId));
-        } else {
-            this.resetTheme();
-        }
+      if (importValue.showDashboardStatistics != null) {
+        this.showDashboardStatistics = importValue.showDashboardStatistics;
+      }
 
-        if (this.localStorage.exists(DBkeys.HOME_URL)) {
-            this._homeUrl = this.localStorage.getDataObject<string>(DBkeys.HOME_URL);
-        }
+      if (importValue.showDashboardNotifications != null) {
+        this.showDashboardNotifications = importValue.showDashboardNotifications;
+      }
 
-        if (this.localStorage.exists(DBkeys.SHOW_DASHBOARD_STATISTICS)) {
-            this._showDashboardStatistics = this.localStorage.getDataObject<boolean>(DBkeys.SHOW_DASHBOARD_STATISTICS);
-        }
+      if (importValue.showDashboardTodo != null) {
+        this.showDashboardTodo = importValue.showDashboardTodo;
+      }
 
-        if (this.localStorage.exists(DBkeys.SHOW_DASHBOARD_NOTIFICATIONS)) {
-            this._showDashboardNotifications = this.localStorage.getDataObject<boolean>(DBkeys.SHOW_DASHBOARD_NOTIFICATIONS);
-        }
-
-        if (this.localStorage.exists(DBkeys.SHOW_DASHBOARD_TODO)) {
-            this._showDashboardTodo = this.localStorage.getDataObject<boolean>(DBkeys.SHOW_DASHBOARD_TODO);
-        }
-
-        if (this.localStorage.exists(DBkeys.SHOW_DASHBOARD_BANNER)) {
-            this._showDashboardBanner = this.localStorage.getDataObject<boolean>(DBkeys.SHOW_DASHBOARD_BANNER);
-        }
+      if (importValue.showDashboardBanner != null) {
+        this.showDashboardBanner = importValue.showDashboardBanner;
+      }
     }
 
-    private saveToLocalStore(data: any, key: string) {
-        setTimeout(() => this.localStorage.savePermanentData(data, key));
+    this.onConfigurationImported.next();
+  }
+
+  public export(changesOnly = true): string {
+    const exportValue: UserConfiguration = {
+      language: changesOnly ? this._language : this.language,
+      themeId: changesOnly ? this._themeId : this.themeId,
+      homeUrl: changesOnly ? this._homeUrl : this.homeUrl,
+      showDashboardStatistics: changesOnly ? this._showDashboardStatistics : this.showDashboardStatistics,
+      showDashboardNotifications: changesOnly ? this._showDashboardNotifications : this.showDashboardNotifications,
+      showDashboardTodo: changesOnly ? this._showDashboardTodo : this.showDashboardTodo,
+      showDashboardBanner: changesOnly ? this._showDashboardBanner : this.showDashboardBanner
+    };
+
+    return JSON.stringify(exportValue);
+  }
+
+  public clearLocalChanges() {
+    this._language = null;
+    this._themeId = null;
+    this._homeUrl = null;
+    this._showDashboardStatistics = null;
+    this._showDashboardNotifications = null;
+    this._showDashboardTodo = null;
+    this._showDashboardBanner = null;
+
+    this.localStorage.deleteData(DBkeys.LANGUAGE);
+    this.localStorage.deleteData(DBkeys.THEME_ID);
+    this.localStorage.deleteData(DBkeys.HOME_URL);
+    this.localStorage.deleteData(DBkeys.SHOW_DASHBOARD_STATISTICS);
+    this.localStorage.deleteData(DBkeys.SHOW_DASHBOARD_NOTIFICATIONS);
+    this.localStorage.deleteData(DBkeys.SHOW_DASHBOARD_TODO);
+    this.localStorage.deleteData(DBkeys.SHOW_DASHBOARD_BANNER);
+
+    this.resetLanguage();
+    this.resetTheme();
+    this.clearUserConfigKeys();
+  }
+
+  private resetLanguage() {
+    const language = this.translationService.useBrowserLanguage();
+
+    if (language) {
+      this._language = language;
+    } else {
+      this._language = this.translationService.useDefaultLangage();
     }
+  }
 
-    public import(jsonValue: string) {
-        this.clearLocalChanges();
+  private resetTheme() {
+    this.themeManager.installTheme();
+    this._themeId = null;
+  }
 
-        if (jsonValue) {
-            const importValue: UserConfiguration = Utilities.JsonTryParse(jsonValue);
 
-            if (importValue.language != null) {
-                this.language = importValue.language;
-            }
+  private addKeyToUserConfigKeys(configKey: string) {
+    const configKeys = this.localStorage.getDataObject<string[]>(DBkeys.USER_CONFIG_KEYS) ?? [];
 
-            if (importValue.themeId != null) {
-                this.themeId = +importValue.themeId;
-            }
-
-            if (importValue.homeUrl != null) {
-                this.homeUrl = importValue.homeUrl;
-            }
-
-            if (importValue.showDashboardStatistics != null) {
-                this.showDashboardStatistics = importValue.showDashboardStatistics;
-            }
-
-            if (importValue.showDashboardNotifications != null) {
-                this.showDashboardNotifications = importValue.showDashboardNotifications;
-            }
-
-            if (importValue.showDashboardTodo != null) {
-                this.showDashboardTodo = importValue.showDashboardTodo;
-            }
-
-            if (importValue.showDashboardBanner != null) {
-                this.showDashboardBanner = importValue.showDashboardBanner;
-            }
-        }
-
-        this.onConfigurationImported.next();
+    if (!configKeys.includes(configKey)) {
+      configKeys.push(configKey);
+      this.localStorage.savePermanentData(configKeys, DBkeys.USER_CONFIG_KEYS);
     }
+  }
 
-    public export(changesOnly = true): string {
-        const exportValue: UserConfiguration = {
-            language: changesOnly ? this._language : this.language,
-            themeId: changesOnly ? this._themeId : this.themeId,
-            homeUrl: changesOnly ? this._homeUrl : this.homeUrl,
-            showDashboardStatistics: changesOnly ? this._showDashboardStatistics : this.showDashboardStatistics,
-            showDashboardNotifications: changesOnly ? this._showDashboardNotifications : this.showDashboardNotifications,
-            showDashboardTodo: changesOnly ? this._showDashboardTodo : this.showDashboardTodo,
-            showDashboardBanner: changesOnly ? this._showDashboardBanner : this.showDashboardBanner
-        };
+  private clearUserConfigKeys() {
+    const configKeys = this.localStorage.getDataObject<string[]>(DBkeys.USER_CONFIG_KEYS);
 
-        return JSON.stringify(exportValue);
+    if (configKeys != null && configKeys.length > 0) {
+      for (let key of configKeys) {
+        this.localStorage.deleteData(key);
+      }
+
+      this.localStorage.deleteData(DBkeys.USER_CONFIG_KEYS);
     }
+  }
 
-    public clearLocalChanges() {
-        this._language = null;
-        this._themeId = null;
-        this._homeUrl = null;
-        this._showDashboardStatistics = null;
-        this._showDashboardNotifications = null;
-        this._showDashboardTodo = null;
-        this._showDashboardBanner = null;
+  public saveConfiguration(data: any, configKey: string) {
+    this.addKeyToUserConfigKeys(configKey);
+    this.localStorage.savePermanentData(data, configKey)
+  }
 
-        this.localStorage.deleteData(DBkeys.LANGUAGE);
-        this.localStorage.deleteData(DBkeys.THEME_ID);
-        this.localStorage.deleteData(DBkeys.HOME_URL);
-        this.localStorage.deleteData(DBkeys.SHOW_DASHBOARD_STATISTICS);
-        this.localStorage.deleteData(DBkeys.SHOW_DASHBOARD_NOTIFICATIONS);
-        this.localStorage.deleteData(DBkeys.SHOW_DASHBOARD_TODO);
-        this.localStorage.deleteData(DBkeys.SHOW_DASHBOARD_BANNER);
-
-        this.resetLanguage();
-        this.resetTheme();
-    }
-
-    private resetLanguage() {
-        const language = this.translationService.useBrowserLanguage();
-
-        if (language) {
-            this._language = language;
-        } else {
-            this._language = this.translationService.useDefaultLangage();
-        }
-    }
-
-    private resetTheme() {
-        this.themeManager.installTheme();
-        this._themeId = null;
-    }
+  public getConfiguration<T>(configKey: string, isDateType = false): T {
+    return this.localStorage.getDataObject<T>(configKey, isDateType);
+  }
 }
