@@ -4,8 +4,7 @@
 // =============================
 
 export const environment = {
-    production: true,
-    baseUrl: null, // Change this to the address of your backend API if different from frontend address
-    tokenUrl: null, // For IdentityServer/Authorization Server API. You can set to null if same as baseUrl
-    loginUrl: '/login'
+  production: true,
+  baseUrl: null, // Change this to the address of your backend API if different from frontend address
+  loginUrl: '/login'
 };
