@@ -4,14 +4,13 @@
 // =============================
 
 // This file can be replaced during build by using the `fileReplacements` array.
-// `ng build --prod` replaces `environment.ts` with `environment.prod.ts`.
+// `ng build` replaces `environment.ts` with `environment.prod.ts`.
 // The list of file replacements can be found in `angular.json`.
 
 export const environment = {
-    production: false,
-    baseUrl: null, // Change this to the address of your backend API if different from frontend address
-    tokenUrl: null, // For IdentityServer/Authorization Server API. You can set to null if same as baseUrl
-    loginUrl: '/login'
+  production: false,
+  baseUrl: null, // Change this to the address of your backend API if different from frontend address
+  loginUrl: '/login'
 };
 
 /*
@@ -21,4 +20,4 @@ export const environment = {
  * This import should be commented out in production mode because it will have a negative impact
  * on performance if an error is thrown.
  */
-// import 'zone.js/dist/zone-error';  // Included with Angular CLI.
+// import 'zone.js/plugins/zone-error';  // Included with Angular CLI.
