@@ -17,37 +17,14 @@ namespace QuickApp.Helpers
 {
     public static class Utilities
     {
-        static ILoggerFactory _loggerFactory;
-
-
-        public static void ConfigureLogger(ILoggerFactory loggerFactory)
+        public static void QuickLog(string text, string logPath)
         {
-            _loggerFactory = loggerFactory;
-        }
-
-
-        public static ILogger CreateLogger<T>()
-        {
-            //Usage: Utilities.CreateLogger<SomeClass>().LogError(LoggingEvents.SomeEventId, ex, "An error occurred because of xyz");
-
-            if (_loggerFactory == null)
-            {
-                throw new InvalidOperationException($"{nameof(ILogger)} is not configured. {nameof(ConfigureLogger)} must be called before use");
-                //_loggerFactory = new LoggerFactory().AddConsole().AddDebug();
-            }
-
-            return _loggerFactory.CreateLogger<T>();
-        }
-
-
-        public static void QuickLog(string text, string filename)
-        {
-            string dirPath = Path.GetDirectoryName(filename);
+            string dirPath = Path.GetDirectoryName(logPath);
 
             if (!Directory.Exists(dirPath))
                 Directory.CreateDirectory(dirPath);
 
-            using (StreamWriter writer = File.AppendText(filename))
+            using (StreamWriter writer = File.AppendText(logPath))
             {
                 writer.WriteLine($"{DateTime.Now} - {text}");
             }
